requirements.txt
# Core dependencies
pandas==2.1.4
numpy==1.24.3
requests==2.31.0
python-dotenv==1.0.0

# Streaming and messaging
kafka-python==2.0.2
confluent-kafka==2.3.0

# Spark and processing
pyspark==3.5.0
delta-spark==3.0.0

# Data quality
great-expectations==0.18.7

# Database
psycopg2-binary==2.9.9
sqlalchemy==2.0.23

# Visualization
streamlit==1.29.0
plotly==5.18.0
keplergl==0.3.2
folium==0.15.1
altair==5.2.0

# Orchestration
apache-airflow==2.8.0
apache-airflow-providers-postgres==5.9.0
apache-airflow-providers-docker==3.9.0

# Weather API
openweather==1.1.0

# Monitoring and alerting
prometheus-client==0.19.0
email-validator==2.1.0
twilio==8.11.0

# Development
pytest==7.4.3
black==23.12.1
flake8==7.0.0
pre-commit==3.6.0

# Utilities
pytz==2023.3
pendulum==3.0.0
pydantic==2.5.3
loguru==0.7.2